LayoutTests/imported/w3c/web-platform-tests/permissions/permissions-query-feature-policy-attribute.https.sub-expected.txt


PASS Permissions.state is "prompt" with allow="geolocation" in same-origin iframes.
PASS Permissions.state is "prompt" with allow="geolocation" in cross-origin iframes.
PASS Permission.state is "prompt" in same-origin iframes.
PASS Permission.state is "denied" in cross-origin iframes.
